+++ src/contract/ContractCreateTransaction.js
@@ -61,6 +61,7 @@ export default class ContractCreateTransaction extends Transaction {
      * @param {Duration | Long | number} [props.autoRenewPeriod]
      * @param {Uint8Array} [props.constructorParameters]
      * @param {string} [props.contractMemo]
+     * @param {number} [props.maxAutomaticTokenAssociations]
      */
     constructor(props = {}) {
         super();
@@ -119,6 +120,12 @@ export default class ContractCreateTransaction extends Transaction {
          */
         this._contractMemo = null;
 
+        /**
+         * @private
+         * @type {number}
+         */
+        this._maxAutomaticTokenAssociations = 0;
+
         this._defaultMaxTransactionFee = new Hbar(20);
 
         if (props.bytecodeFileId != null) {
@@ -155,6 +162,12 @@ export default class ContractCreateTransaction extends Transaction {
 
         if (props.contractMemo != null) {
             this.setContractMemo(props.contractMemo);
+        }
+
+        if (props.maxAutomaticTokenAssociations != null) {
+            this.setMaxAutomaticTokenAssociations(
+                props.maxAutomaticTokenAssociations
+            );
         }
     }
 
@@ -218,6 +231,10 @@ export default class ContractCreateTransaction extends Transaction {
                         ? create.constructorParameters
                         : undefined,
                 contractMemo: create.memo != null ? create.memo : undefined,
+                maxAutomaticTokenAssociations:
+                    create.maxAutomaticTokenAssociations != null
+                        ? create.maxAutomaticTokenAssociations
+                        : undefined,
             }),
             transactions,
             signedTransactions,
@@ -409,6 +426,23 @@ export default class ContractCreateTransaction extends Transaction {
     }
 
     /**
+     * @returns {?number}
+     */
+    get maxAutomaticTokenAssociations() {
+        return this._maxAutomaticTokenAssociations;
+    }
+
+    /**
+     * @param {number} maxAutomaticTokenAssociations
+     * @returns {this}
+     */
+    setMaxAutomaticTokenAssociations(maxAutomaticTokenAssociations) {
+        this._maxAutomaticTokenAssociations = maxAutomaticTokenAssociations;
+
+        return this;
+    }
+
+    /**
      * @param {Client} client
      */
     _validateChecksums(client) {
@@ -467,6 +501,7 @@ export default class ContractCreateTransaction extends Transaction {
             autoRenewPeriod: this._autoRenewPeriod._toProtobuf(),
             constructorParameters: this._constructorParameters,
             memo: this._contractMemo,
+            maxAutomaticTokenAssociations: this._maxAutomaticTokenAssociations,
         };
     }