+++ src/contract/ContractUpdateTransaction.js
@@ -57,6 +57,7 @@ export default class ContractUpdateTransaction extends Transaction {
      * @param {AccountId | string} [props.proxyAccountId]
      * @param {Duration | Long | number} [props.autoRenewPeriod]
      * @param {string} [props.contractMemo]
+     * @param {number} [props.maxAutomaticTokenAssociations]
      */
     constructor(props = {}) {
         super();
@@ -103,6 +104,12 @@ export default class ContractUpdateTransaction extends Transaction {
          */
         this._contractMemo = null;
 
+        /**
+         * @private
+         * @type {?number}
+         */
+        this._maxAutomaticTokenAssociations = 0;
+
         if (props.contractId != null) {
             this.setContractId(props.contractId);
         }
@@ -129,6 +136,12 @@ export default class ContractUpdateTransaction extends Transaction {
 
         if (props.contractMemo != null) {
             this.setContractMemo(props.contractMemo);
+        }
+
+        if (props.maxAutomaticTokenAssociations != null) {
+            this.setMaxAutomaticTokenAssociations(
+                props.maxAutomaticTokenAssociations
+            );
         }
     }
 
@@ -167,6 +180,15 @@ export default class ContractUpdateTransaction extends Transaction {
             contractMemo = update.memoWrapper.value;
         }
 
+        let maxAutomaticTokenAssociations = undefined;
+        if (
+            update.maxAutomaticTokenAssociations != null &&
+            update.maxAutomaticTokenAssociations.value != null
+        ) {
+            maxAutomaticTokenAssociations =
+                update.maxAutomaticTokenAssociations.value;
+        }
+
         return Transaction._fromProtobufTransactions(
             new ContractUpdateTransaction({
                 contractId:
@@ -203,6 +225,7 @@ export default class ContractUpdateTransaction extends Transaction {
                         : undefined,
                 autoRenewPeriod,
                 contractMemo,
+                maxAutomaticTokenAssociations,
             }),
             transactions,
             signedTransactions,
@@ -369,6 +392,24 @@ export default class ContractUpdateTransaction extends Transaction {
     }
 
     /**
+     * @returns {number | null}
+     */
+    get maxAutomaticTokenAssociations() {
+        return this._maxAutomaticTokenAssociations;
+    }
+
+    /**
+     * @param {number} maxAutomaticTokenAssociations
+     * @returns {this}
+     */
+    setMaxAutomaticTokenAssociations(maxAutomaticTokenAssociations) {
+        this._requireNotFrozen();
+        this._maxAutomaticTokenAssociations = maxAutomaticTokenAssociations;
+
+        return this;
+    }
+
+    /**
      * @param {Client} client
      */
     _validateChecksums(client) {
@@ -437,6 +478,12 @@ export default class ContractUpdateTransaction extends Transaction {
                 this._contractMemo != null
                     ? {
                           value: this._contractMemo,
+                      }
+                    : null,
+            maxAutomaticTokenAssociations:
+                this._maxAutomaticTokenAssociations != null
+                    ? {
+                          value: this._maxAutomaticTokenAssociations,
                       }
                     : null,
         };